Put in common gnn features extractor between ray.rllib and sb3 wrappers
The GNN features extractor used by rallib and sb3 are the same (except
the one for sb3 derives from a BaseFeaturesExtractor) and so is the
function conveting GraphInstance into torch_geometric.data.Data.

We put that in common in hub/solver/utils/gnn.

We also remove unused code in ray.rllib gnn code.
